refactor : delete package product under dto
