+++ customer-api/src/main/java/zerobase/customerapi/service/CustomerProductService.java
@@ -12,8 +12,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import zerobase.common.dto.product.ProductDto;
-import zerobase.common.dto.product.ProductListDto;
+import zerobase.common.dto.ProductDto;
+import zerobase.common.dto.ProductListDto;
 import zerobase.common.entity.ProductEntity;
 import zerobase.common.exception.CommonCustomException;
 import zerobase.common.repository.ProductRepository;
