+++ seller-api/src/main/java/zerobase/sellerapi/service/SellerProductService.java
@@ -7,9 +7,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import zerobase.common.dto.product.ProductDto;
-import zerobase.common.dto.product.ProductEditDto;
-import zerobase.common.dto.product.RegistrationDto;
+import zerobase.common.dto.ProductDto;
+import zerobase.common.dto.ProductEditDto;
+import zerobase.common.dto.RegistrationDto;
 import zerobase.common.entity.ProductEntity;
 import zerobase.common.exception.CommonCustomException;
 import zerobase.common.repository.ProductRepository;
